redirect to the event page after login successfully
