+++ .vscode/settings.json
@@ -1,3 +1,5 @@
 {
-    "editor.acceptSuggestionOnEnter": "on"
+    "editor.acceptSuggestionOnEnter": "on",
+    "pieces.cloudCapabilities": "Blended",
+    "pieces.telemetry": true
 }
