+++ src/app/events/page.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-hooks/rules-of-hooks */
 "use client";
 import { ClientEventCard } from "@/components/events/ClientEventCard";
 import React, { useEffect, useState } from "react";
@@ -18,7 +19,10 @@ const page = () => {
     <div className="mt-16 px-6 py-3">
       <h1 className="mb-6 text-xl font-bold">Events</h1>
       <div className="grid grid-cols-3 gap-4 ">
-        {events && events.map((event: any) => <ClientEventCard key={event.id} event={event} />)}
+        {events &&
+          events.map((event: any) => (
+            <ClientEventCard key={event.id} event={event} />
+          ))}
       </div>
     </div>
   );
